updatesEnabled
Validatables now have an updatesEnabled value. What this does is is indicate if input and validation are disabled or not. if disabled, the user will not be able to input into the Fields and cannot validate anything. This functionality has been added so uploading to a server in the future can occur.

- New Form.Field.prefilledDisabled, for creating prefilled fields that cannot be used by the user.
- New Form.Validatable.possiblyToggleErr, a convenience function.
- Some better readme stuff.